+++ xerparser/schemas/account.py
@@ -1,15 +1,11 @@
 # xerparser
 # account.py
 
-from pydantic import BaseModel, Field, validator
 
-
-class ACCOUNT(BaseModel):
+class ACCOUNT:
     """
     A class to represent a cost account.
-
     ...
-
     Attributes
     ----------
     uid: str
@@ -20,20 +16,20 @@ class ACCOUNT(BaseModel):
         Cost Account Description [acct_desc]
     name: str
         Cost Account Name [acct_name]
-
     """
 
-    uid: str = Field(alias="acct_id")
-    code: str = Field(alias="acct_short_name")
-    description: str | None
-    name: str = Field(alias="acct_name")
-
-    @validator("description", pre=True)
-    def empty_str_to_none(cls, value):
-        return (value, None)[value == "" or value == "ï»¿"]
+    def __init__(self, **kwargs) -> None:
+        self.uid: str = kwargs["acct_id"]
+        self.code: str = kwargs["acct_short_name"]
+        self.description: str = _check_description(kwargs["acct_desc"])
+        self.name: str = kwargs["acct_name"]
 
     def __eq__(self, __o: "ACCOUNT") -> bool:
         return self.name == __o.name and self.code == __o.code
 
     def __hash__(self) -> int:
         return hash((self.name, self.name))
+
+
+def _check_description(value: str) -> str:
+    return (value, "")[value == "" or value == "ï»¿"]